restaurant_project/requirements.txt
asgiref==3.8.1
boto3==1.38.44
botocore==1.38.44
crispy-bootstrap5==0.7
Django==4.2.7
django-cleanup==8.0.0
django-crispy-forms==2.0
django-jazzmin==3.0.1
django-storages==1.14.6
jmespath==1.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==1.26.20
